Rewrite flipper packet management

+++ esp32/main/common.c
@@ -353,7 +353,7 @@ void repeat_bytes(uint8_t byte, uint8_t count) {
 }
 
 void send_bytes(uint8_t *bytes, uint8_t size) {
-    for (int i = 0; i < size; ++i) {
+    for (uint8_t i = 0; i < size; ++i) {
         putc(bytes[i], stdout);
     }
     fflush(stdout);
